+++ packages/backend/src/scheduler/SchedulerTask.test.ts
@@ -1,0 +1,41 @@
+import SchedulerTask from './SchedulerTask';
+import {
+    resultsWithOneRow,
+    resultsWithTwoRows,
+    thresholdIncreasedByMock,
+    thresholdLessThanMock,
+} from './SchedulerTask.mock';
+
+describe('isPositiveThresholdAlert', () => {
+    it('should return false if there are no results or no thresholds', () => {
+        expect(
+            SchedulerTask.isPositiveThresholdAlert([thresholdLessThanMock], []),
+        ).toBe(false);
+
+        expect(
+            SchedulerTask.isPositiveThresholdAlert([], resultsWithOneRow),
+        ).toBe(false);
+    });
+    it('should return false if operation requires second row but there isnt one', () => {
+        expect(
+            SchedulerTask.isPositiveThresholdAlert(
+                [thresholdIncreasedByMock],
+                resultsWithOneRow,
+            ),
+        ).toBe(false);
+    });
+    it('should return true if condition match', () => {
+        expect(
+            SchedulerTask.isPositiveThresholdAlert(
+                [thresholdLessThanMock],
+                resultsWithOneRow,
+            ),
+        ).toBe(true);
+        expect(
+            SchedulerTask.isPositiveThresholdAlert(
+                [thresholdIncreasedByMock],
+                resultsWithTwoRows,
+            ),
+        ).toBe(true);
+    });
+});